pyproject.toml
[tool.poetry]
name = "ethglobal2024"
version = "0.1.0"
description = ""
authors = ["George <george.neokq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
web3 = "^7.2.0"
python-telegram-bot = "^21.6"
eth-account = "^0.13.3"
mysql-connector-python = "^9.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
